[Refactor] `ExportMap`: extract "builder" logic to separate files
