+++ src/rules/export.js
@@ -1,4 +1,5 @@
-import ExportMapBuilder, { recursivePatternCapture } from '../exportMapBuilder';
+import ExportMapBuilder from '../exportMap/builder';
+import recursivePatternCapture from '../exportMap/patternCapture';
 import docsUrl from '../docsUrl';
 import includes from 'array-includes';
 import flatMap from 'array.prototype.flatmap';
