+++ src/rules/named.js
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import ExportMapBuilder from '../exportMapBuilder';
+import ExportMapBuilder from '../exportMap/builder';
 import docsUrl from '../docsUrl';
 
 module.exports = {
