+++ src/rules/no-named-as-default-member.js
@@ -4,7 +4,7 @@
  * @copyright 2016 Desmond Brand. All rights reserved.
  * See LICENSE in root directory for full license.
  */
-import ExportMapBuilder from '../exportMapBuilder';
+import ExportMapBuilder from '../exportMap/builder';
 import importDeclaration from '../importDeclaration';
 import docsUrl from '../docsUrl';
