+++ src/rules/no-unused-modules.js
@@ -13,7 +13,8 @@ import values from 'object.values';
 import includes from 'array-includes';
 import flatMap from 'array.prototype.flatmap';
 
-import ExportMapBuilder, { recursivePatternCapture } from '../exportMapBuilder';
+import ExportMapBuilder from '../exportMap/builder';
+import recursivePatternCapture from '../exportMap/patternCapture';
 import docsUrl from '../docsUrl';
 
 let FileEnumerator;
